requirements.txt
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0